requirements.txt
Django
psycopg2
django-environ==0.4.4
djangorestframework
markdown
django-filter
django-rest-auth[with_social]